refactoring data_manager
